Enforce hermitian symmetry for real fields
All fields are complex but some (psi and w) should have imaginary part equal to zero. In dedalus the imaginary part can drift without bound and so we send the fields to grid space every 1000 iterations to prevent this. Not sure if this is strictly necessary when the fields are being simulated as complex?